+++ src/services/post.service.ts
@@ -1,21 +1,25 @@
 import { prisma } from '../config/prisma';
 import { HttpException } from '../exceptions/httpException';
 
-const getTopicPosts = async (animalTypeId: string, topicId: string) => {
-  // const posts = await prisma.post.findMany({
-  //   where: { ownerTopicId: id },
-  //   select: {
-  //     id: true,
-  //     title: true,
-  //     content: true,
-  //     ownerTopic: true,
-  //     ownerUser: true,
-  //   },
-  const topicPosts = await prisma.$queryRaw`select p.*
-                                            from "Post" p
-                                            join "Topic" t on t.id = ${topicId}
-                                            join "AnimalType" a on a.id = ${animalTypeId}
-                                            where p."ownerTopicId" = t.id`;
+const getTopicPost = async (animalTypeId: string, topicId: string) => {
+  const animalType = await prisma.animalType.findUnique({ where: { id: animalTypeId } });
+  if (!animalType) {
+    throw new HttpException(404, 'Animal type not found!');
+  }
+  const topic = await prisma.topic.findUnique({ where: { id: topicId } });
+  if (!topic) {
+    throw new HttpException(404, 'Topic not found!');
+  }
+  const topicPosts = await prisma.post.findMany({
+    where: { ownerTopicId: topicId },
+    select: {
+      id: true,
+      title: true,
+      content: true,
+      ownerTopicId: true,
+      ownerUserId: true,
+    },
+  });
   return topicPosts;
 };
 
@@ -33,7 +37,15 @@ const getPosts = async () => {
   return posts;
 };
 
-const getPost = async (id: string) => {
+const getPost = async (animalTypeId: string, topicId: string, id: string) => {
+  const animalType = await prisma.animalType.findUnique({ where: { id: animalTypeId } });
+  if (!animalType) {
+    throw new HttpException(404, 'Animal type not found!');
+  }
+  const topic = await prisma.topic.findUnique({ where: { id: topicId } });
+  if (!topic) {
+    throw new HttpException(404, 'Topic not found!');
+  }
   const post = await prisma.post.findUnique({ where: { id } });
   if (!post) {
     throw new HttpException(404, 'Post not found!');
@@ -41,24 +53,52 @@ const getPost = async (id: string) => {
   return post;
 };
 
-const createPost = async (data: {
-  title: string;
-  content: string;
-  ownerUserId: string;
-  ownerTopicId: string;
-}) => {
+const createPost = async (
+  animalTypeId: string,
+  topicId: string,
+  data: {
+    title: string;
+    content: string;
+    ownerUserId: string;
+  }
+) => {
+  const animalType = await prisma.animalType.findUnique({ where: { id: animalTypeId } });
+  if (!animalType) {
+    throw new HttpException(404, 'Animal type not found!');
+  }
+  const topic = await prisma.topic.findUnique({ where: { id: topicId } });
+  if (!topic) {
+    throw new HttpException(404, 'Topic not found!');
+  }
   const post = await prisma.post.create({
     data: {
       title: data.title,
       content: data.content,
       ownerUserId: data.ownerUserId,
-      ownerTopicId: data.ownerTopicId,
+      ownerTopicId: topicId,
     },
   });
   return post;
 };
 
-const updatePost = async (data: { title: string; content: string }, id: string) => {
+const updatePost = async (
+  animalTypeId: string,
+  topicId: string,
+  id: string,
+  data: { title: string; content: string }
+) => {
+  const animalType = await prisma.animalType.findUnique({ where: { id: animalTypeId } });
+  if (!animalType) {
+    throw new HttpException(404, 'Animal type not found!');
+  }
+  const topic = await prisma.topic.findUnique({ where: { id: topicId } });
+  if (!topic) {
+    throw new HttpException(404, 'Topic not found!');
+  }
+  const tempPost = await prisma.post.findUnique({ where: { id } });
+  if (!tempPost) {
+    throw new HttpException(404, 'Post not found!');
+  }
   const post = await prisma.post.update({
     where: { id: id },
     data: { title: data.title, content: data.content },
@@ -66,15 +106,27 @@ const updatePost = async (data: { title: string; content: string }, id: string) 
   return post;
 };
 
-const deletePost = async (id: string) => {
+const deletePost = async (animalTypeId: string, topicId: string, id: string) => {
+  const animalType = await prisma.animalType.findUnique({ where: { id: animalTypeId } });
+  if (!animalType) {
+    throw new HttpException(404, 'Animal type not found!');
+  }
+  const topic = await prisma.topic.findUnique({ where: { id: topicId } });
+  if (!topic) {
+    throw new HttpException(404, 'Topic not found!');
+  }
+  const tempPost = await prisma.post.findUnique({ where: { id } });
+  if (!tempPost) {
+    throw new HttpException(404, 'Post not found!');
+  }
   const post = await prisma.post.delete({
     where: { id: id },
   });
-  return post;
+  // return post;
 };
 
 export const postService = {
-  getTopicPosts,
+  getTopicPost,
   getPosts,
   getPost,
   createPost,
